requirements.txt
python-dotenv==1.0.1
notion-client==2.2.1
requests==2.32.3
beautifulsoup4==4.12.3
